+++ src/Dynamics365Configuration/Install-Dynamics365ReportingExtensions.psm1
@@ -90,12 +90,18 @@
             Write-Output "$(Get-Date) Started installation job, log will be saved in $logFilePath";
             $lastLinesCount = 0;
             $startTime = Get-Date;
+            $logReadIterationsLeft = 2;
             Start-Sleep $logFilePullIntervalInSeconds;
             do {
                 $elapsedTime = $( Get-Date ) - $startTime;
                 $elapsedString = "{0:HH:mm:ss}" -f ( [datetime]$elapsedTime.Ticks );
                 Write-Output "$(Get-Date) Elapsed $elapsedString. Waiting until CRM reporting extensions job is done, sleeping $logFilePullIntervalInSeconds sec";
                 Start-Sleep $logFilePullIntervalInSeconds;
+                if ( $job.State -eq "Completed" ) {
+                    Write-Output "$(Get-Date) Log read iterations left: $logReadIterationsLeft"
+                    $logReadIterationsLeft--;
+                    Start-Sleep 30;
+                }
                 if(($logFilePullToOutput -eq $True) -and ((Test-Path $logFilePath) -eq $True)) {
 
                     $linesCount    = (Get-Content $logFilePath | Measure-Object -Line).Lines;
@@ -109,12 +115,12 @@
                             Write-Output $line;
                         }
                     } else {
-                       Write-Output "$(Get-Date) - no new logs";
+                        Write-Output "$(Get-Date) - no new logs";
                     }
 
                     $lastLinesCount = $linesCount;
                 }
-            } while ( $job.State -ne "Completed" )
+            } until ( ( $job.State -eq "Completed" ) -and ( $logReadIterationsLeft -le 0 ) )
             $elapsedTime = $( Get-Date ) - $startTime;
             $elapsedString = "{0:HH:mm:ss}" -f ( [datetime]$elapsedTime.Ticks );
             Write-Output "$(Get-Date) Elapsed $elapsedString. Job is complete, output:";